command: Add SET_CLOSURE_AND_ORIENTATION
Add SET_CLOSURE_AND_ORIENTATION to the command enum. This command
is supported by Somfy J4 io Protect motors and allows to set the blind
closure as well as the slide orientation in one command.

The main advantage for using this command (compared to individual
commands) is that the slides will not change the orientation if the
motor is already in the correct closure position. There's also no need
to wait for the target position to be reached before modifying the
orientation. It's also helpful for automating multiple motors, because
it allows that in less commands overall.